raise OSError if tomcat stop fails
git-svn-id: https://dev.lagrummet.se/svn/rinfo/trunk@1360 f32dfbcc-ef79-4080-89f5-1451246e9e1e

+++ manage/deploy/__init__.py
@@ -1,4 +1,5 @@
 from fabric.api import *
+from exceptions import OSError
 from util import venv
 from targetenvs import _needs_targetenv
 
@@ -9,7 +10,9 @@ def local_lib_rinfo_pkg():
 def _deploy_war(localwar, warname):
     _needs_targetenv()
     put(localwar, "%(dist_dir)s/%(warname)s.war"%venv())
-    sudo("%(tomcat_stop)s"%venv())
+    result = sudo("%(tomcat_stop)s"%venv())
+    if result.failed:
+        raise OSError(result)
     sudo("rm -rf %(tomcat_webapps)s/%(warname)s/"%venv())
     sudo("mv %(dist_dir)s/%(warname)s.war %(tomcat_webapps)s/"%venv())
     sudo("chown %(tomcat_user)s %(tomcat_webapps)s/%(warname)s.war"%venv())
